feat: 모든 ScreenUtilService 제거 완료
